fix: reswizzle DocSidebarItem
