+++ src/theme/DocSidebarItem/Link.module.css
@@ -1,3 +1,7 @@
+.menuExternalLink {
+  align-items: center;
+}
+
 /**
  * Copyright (c) Facebook, Inc. and its affiliates.
  *
